pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bank-complaint-audit-issues"
version = "0.1.0"
description = "Audit-oriented complaint issue discovery and classification engine."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "taotiantian8666666" }
]
dependencies = [
  "pandas>=2.2,<3.0",
  "pyarrow>=16.0,<19.0",
  "PyYAML>=6.0,<7.0",
  "pydantic>=2.8,<3.0",
  "typer>=0.12,<1.0",
  "rich>=13.7,<14.0",
  "duckdb>=1.0,<2.0",
  "jieba>=0.42,<1.0",
  "RapidFuzz>=3.9,<4.0",
  "regex>=2024.5.15",
  "numpy>=1.26,<3.0",
  "scipy>=1.13,<2.0",
  "scikit-learn>=1.5,<2.0",
  "joblib>=1.4,<2.0",
  "openpyxl>=3.1,<4.0",
]

[project.scripts]
audit-issues = "audit_issue_engine.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

